Keep URLs out of refresh errors and make the counters concurrency-safe
The refresh path redacted URLs from its structured warnings but still
interpolated them into the ValueError and URLError it raises, so a
credential-bearing URL — userinfo or a query token — could reach a traceback or
any caller doing str(exc). Report the offending scheme instead; the tests match
on the surviving wording.

Guard REFRESH_DEGRADATIONS with a process-wide lock: incrementing a dict value
is a read-modify-write, so concurrent refreshes could lose increments. Add a
reset hook and a snapshot accessor so tests own their starting state instead of
asserting on deltas, and cover both concurrent increments and reset.

Split the 513-line rollout test module, which exceeded the 400-line ceiling,
into dictionary/render and refresh/redirect-policy suites with the shared
import scaffolding in a conftest, and collect both from the Makefile target.
