+++ pyproject.toml
@@ -124,6 +124,7 @@ extend-ignore = [
 # allowances as the tests they support; they use bare ``assert`` for fixtures.
 "cuprum/unittests/*_test_helpers.py" = ["S101", "PLR0913", "PLR0917", "PLR2004", "PLR6301"]
 "cuprum/unittests/*_support.py" = ["S101", "PLR0913", "PLR0917", "PLR2004", "PLR6301"]
+"**/conftest.py" = ["S101", "PLR0913", "PLR0917", "PLR2004", "PLR6301"]
 "tests/steps/*.py" = ["S101", "PLR0913", "PLR0917", "PLR2004", "PLR6301"]
 
 [tool.ruff.lint.flake8-import-conventions]
